use (signed) int in for-loop for compatiblity with MSVC's OpenMP 2.0
in MSVC OpenMP the for-loop index can't be unsigned

https://msdn.microsoft.com/en-us/library/98affhcy.aspx
error C3016: 'i' : index variable in OpenMP 'for' statement must have signed integral type

Also I define `charstring_pool_t.count` as uint16_t to avoid warnings when comparing it with the signed int for-loop index.

In the CFF spec, the `count` attribute is defined as a `Card16`, or 2-byte unsigned number (i.e. unit16_t).

+++ cxx-src/cffCompressor.h
@@ -196,7 +196,7 @@ class charstring_pool_t {
     std::vector<uint8_t> fdSelect;
     std::vector<unsigned> rev;
     bool fdSelectTrivial;
-    unsigned count;
+    uint16_t count;
     bool finalized;
     int numRounds;
